+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Feature
 - \[DWD Obs\] Use utf8 encoding for parsing data
 - Add `if_exists` argument to `to_target`
+- Use more polars-native methods
 
 ## 0.113.0 - 2025-09-21